+++ R/praatScriptCentreOfGravity.R
@@ -15,20 +15,20 @@
 #' @seealso \link{praatScriptFormants}
 #' @seealso \link{praatScriptIntensity}
 #' @seealso \link{praatScriptPitch}
+#' @seealso \link{praatScriptFastTrack}
 #' @examples
 #' \dontrun{
 #' ## define the LaBB-CAT URL
 #' labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"
 #' 
 #' ## Perform a search
-#' results <- getMatches(labbcat.url, list(segments="I"))
+#' results <- getMatches(labbcat.url, list(segment="I"))
 #' 
 #' ## Get centres of gravity for all matches
 #' cog <- processWithPraat(
 #'               labbcat.url,
-#'               results$MatchId, results$Target.segments.start, results$Target.segments.end,
-#'               praatScriptCentreOfGravity(powers=c(1.0,2.0)),
-#'               no.progress=TRUE)
+#'               results$MatchId, results$Target.segment.start, results$Target.segment.end,
+#'               praatScriptCentreOfGravity(powers=c(1.0,2.0)))
 #' }
 #' @keywords praat
 #'