+++ R/praatScriptFormants.R
@@ -3,6 +3,9 @@
 #' This function generates a Praat script fragment which can be passed as the praat.script
 #' parameter of \link{processWithPraat}, in order to extract selected formants.
 #'
+#' The \link{praatScriptFastTrack} function provides an alternative to this function which
+#' uses the FastTrack Praat plugin for formant analysis.
+#'
 #' @param formants A vector of integers specifying which formants to extract, e.g c(1,2)
 #'     for the first and second formant.
 #' @param sample.points A vector of numbers (0 <= sample.points <= 1) specifying multiple
@@ -29,19 +32,20 @@
 #' @seealso \link{praatScriptCentreOfGravity}
 #' @seealso \link{praatScriptIntensity}
 #' @seealso \link{praatScriptPitch}
+#' @seealso \link{praatScriptFastTrack}
 #' @examples
 #' \dontrun{
 #' ## define the LaBB-CAT URL
 #' labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"
 #' 
 #' ## Get all tokens of the KIT vowel
-#' results <- getMatches(labbcat.url, list(segments="I"))
+#' results <- getMatches(labbcat.url, list(segment="I"))
 #' 
 #' ## Get the first 3 formants at three points during the vowel
 #' formants <- processWithPraat(
 #'               labbcat.url,
-#'               results$MatchId, results$Target.segments.start, results$Target.segments.end,
-#'               window.offset=0.5,
+#'               results$MatchId, results$Target.segment.start, results$Target.segment.end,
+#'               window.offset=0.025,
 #'               praatScriptFormants(formants=c(1,2,3),
 #'               sample.points=c(0.25,0.5,0.75)))
 #' }