+++ R/praatScriptIntensity.R
@@ -13,20 +13,20 @@
 #' @seealso \link{praatScriptFormants}
 #' @seealso \link{praatScriptCentreOfGravity}
 #' @seealso \link{praatScriptPitch}
+#' @seealso \link{praatScriptFastTrack}
 #' @examples
 #' \dontrun{
 #' ## define the LaBB-CAT URL
 #' labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"
 #' 
 #' ## Perform a search
-#' results <- getMatches(labbcat.url, list(segments="s"))
+#' results <- getMatches(labbcat.url, list(segment="s"))
 #' 
 #' ## Get intensity for all matches
 #' intensity <- processWithPraat(
 #'               labbcat.url,
-#'               results$MatchId, results$Target.segments.start, results$Target.segments.end,
-#'               praatScriptIntensity(),
-#'               no.progress=TRUE)
+#'               results$MatchId, results$Target.segment.start, results$Target.segment.end,
+#'               praatScriptIntensity())
 #' }
 #' @keywords praat
 #'