+++ R/praatScriptPitch.R
@@ -34,20 +34,20 @@
 #' @seealso \link{praatScriptFormants}
 #' @seealso \link{praatScriptCentreOfGravity}
 #' @seealso \link{praatScriptIntensity}
+#' @seealso \link{praatScriptFastTrack}
 #' @examples
 #' \dontrun{
 #' ## define the LaBB-CAT URL
 #' labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"
 #' 
 #' ## Perform a search
-#' results <- getMatches(labbcat.url, list(segments="I"))
+#' results <- getMatches(labbcat.url, list(segment="I"))
 #' 
 #' ## Get pitch mean, max, and min for all matches
 #' pitch <- processWithPraat(
 #'               labbcat.url,
-#'               results$MatchId, results$Target.segments.start, results$Target.segments.end,
-#'               praatScriptPitch(get.mean=TRUE, get.minimum=TRUE, get.maximum=TRUE),
-#'               no.progress=TRUE)
+#'               results$MatchId, results$Target.segment.start, results$Target.segment.end,
+#'               praatScriptPitch(get.mean=TRUE, get.minimum=TRUE, get.maximum=TRUE))
 #' }
 #' @keywords praat
 #'