+++ tests/testthat/test-getMatches.R
@@ -11,7 +11,7 @@ test_that("getMatches works with 1x1 orthographic search using full structure", 
                      orthography = list(pattern = "knox")))))
 
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
 
     ## check dataframe columns
     expect_equal(length(matches$MatchId), 1)
@@ -20,14 +20,14 @@ test_that("getMatches works with 1x1 orthographic search using full structure", 
     expect_equal(length(matches$LineEnd), 1)
     expect_equal(length(matches$Text), 1)
     expect_equal(length(matches$URL), 1)
-    expect_equal(length(matches$Target.transcript), 1)
-    expect_equal(length(matches$Target.transcript.start), 1)
-    expect_equal(length(matches$Target.transcript.end), 1)
+    expect_equal(length(matches$Target.word), 1)
+    expect_equal(length(matches$Target.word.start), 1)
+    expect_equal(length(matches$Target.word.end), 1)
 
     ## search doens't include segments - ensure there's no segment information returned
-    expect_true(is.null(matches$Target.segments))
-    expect_true(is.null(matches$Target.segments.start))
-    expect_true(is.null(matches$Target.segments.end))
+    expect_true(is.null(matches$Target.segment))
+    expect_true(is.null(matches$Target.segment.start))
+    expect_true(is.null(matches$Target.segment.end))
 
     ## check dataframe column types
     expect_false(is.numeric(matches$MatchId))
@@ -36,9 +36,9 @@ test_that("getMatches works with 1x1 orthographic search using full structure", 
     expect_true(is.numeric(matches$LineEnd))
     expect_false(is.numeric(matches$Text))
     expect_false(is.numeric(matches$URL))
-    expect_false(is.numeric(matches$Target.transcript))
-    expect_true(is.numeric(matches$Target.transcript.start))
-    expect_true(is.numeric(matches$Target.transcript.end))
+    expect_false(is.numeric(matches$Target.word))
+    expect_true(is.numeric(matches$Target.word.start))
+    expect_true(is.numeric(matches$Target.word.end))
 
     expect_equal(as.vector(matches$Text)[[1]], "Knox")
 })
@@ -57,7 +57,7 @@ test_that("getMatches works with 2x1 orthographic search using full structure", 
                      orthography = list(not = TRUE, pattern = "xxx")))))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
     
     expect_equal(length(matches$MatchId), 1)
 })
@@ -70,7 +70,7 @@ test_that("getMatches works with 1x1 orthographic search using simple structure"
     pattern <- list(orthography = "knox")
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
 
     expect_equal(length(matches$MatchId), 1)
     expect_equal(as.vector(matches$Text)[[1]], "Knox")
@@ -84,7 +84,7 @@ test_that("getMatches works with 2x1 orthographic search using simple structure"
     pattern <- list(list(orthography = "knox"), list(orthography = "church"))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
 
     expect_equal(length(matches$MatchId), 1)
     expect_equal(as.vector(matches$Text)[[1]], "Knox Church .")
@@ -105,7 +105,7 @@ test_that("getMatches works with 2x3 non-orthographic search using full structur
                      orthography = list(pattern = "h.*")))))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
     
     ## check dataframe columns
     expect_equal(length(matches$MatchId), 1)
@@ -113,9 +113,9 @@ test_that("getMatches works with 2x3 non-orthographic search using full structur
     expect_equal(length(matches$Line), 1)
     expect_equal(length(matches$LineEnd), 1)
     expect_equal(length(matches$Text), 1)
-    expect_equal(length(matches$Target.transcript), 1)
-    expect_equal(length(matches$Target.transcript.start), 1)
-    expect_equal(length(matches$Target.transcript.end), 1)
+    expect_equal(length(matches$Target.word), 1)
+    expect_equal(length(matches$Target.word.start), 1)
+    expect_equal(length(matches$Target.word.end), 1)
 
     ## check dataframe column types
     expect_false(is.numeric(matches$MatchId))
@@ -123,9 +123,9 @@ test_that("getMatches works with 2x3 non-orthographic search using full structur
     expect_true(is.numeric(matches$Line))
     expect_true(is.numeric(matches$LineEnd))
     expect_false(is.numeric(matches$Text))
-    expect_false(is.numeric(matches$Target.transcript))
-    expect_true(is.numeric(matches$Target.transcript.start))
-    expect_true(is.numeric(matches$Target.transcript.end))
+    expect_false(is.numeric(matches$Target.word))
+    expect_true(is.numeric(matches$Target.word.start))
+    expect_true(is.numeric(matches$Target.word.end))
 
     expect_equal(as.vector(matches$Text)[[1]], "Knox Church . had")
 })
@@ -141,7 +141,7 @@ test_that("getMatches works with 2x3 non-orthographic search using simple struct
         list(orthography = "ch.*"))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
 
     expect_equal(as.vector(matches$Text)[[1]], "Knox Church .")
 })
@@ -163,7 +163,7 @@ test_that("getMatches works with complex, multi-match searches", {
                      frequency = list(max = "2")))))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern)
     
     ## check dataframe columns
     expect_true(length(matches$MatchId) >= 1400)
@@ -171,9 +171,9 @@ test_that("getMatches works with complex, multi-match searches", {
     expect_true(length(matches$Line)>= 1400)
     expect_true(length(matches$LineEnd)>= 1400)
     expect_true(length(matches$Text)>= 1400)
-    expect_true(length(matches$Target.transcript)>= 1400)
-    expect_true(length(matches$Target.transcript.start)>= 1400)
-    expect_true(length(matches$Target.transcript.end)>= 1400)
+    expect_true(length(matches$Target.word)>= 1400)
+    expect_true(length(matches$Target.word.start)>= 1400)
+    expect_true(length(matches$Target.word.end)>= 1400)
 
     ## check dataframe column types
     expect_false(is.numeric(matches$MatchId))
@@ -181,12 +181,12 @@ test_that("getMatches works with complex, multi-match searches", {
     expect_true(is.numeric(matches$Line))
     expect_true(is.numeric(matches$LineEnd))
     expect_false(is.numeric(matches$Text))
-    expect_false(is.numeric(matches$Target.transcript))
-    expect_true(is.numeric(matches$Target.transcript.start))
-    expect_true(is.numeric(matches$Target.transcript.end))
+    expect_false(is.numeric(matches$Target.word))
+    expect_true(is.numeric(matches$Target.word.start))
+    expect_true(is.numeric(matches$Target.word.end))
 })
 
-test_that("getMatches includes segment info when segments layer searched", {
+test_that("getMatches includes segment info when segment layer searched", {
     skip_on_cran() # don't run tests that depend on external resource on CRAN
     if (!is.null(labbcatCredentials(labbcat.url, "demo", "demo"))) skip("Server not available")
 
@@ -194,11 +194,10 @@ test_that("getMatches includes segment info when segments layer searched", {
     pattern <- list(
         columns = list(
             list(layers = list(
-                     segments = list(pattern = "I")))))
+                     segment = list(pattern = "I")))))
     
     ## get matches
-    matches <- getMatches(labbcat.url, pattern, participant.ids="UC427_ViktoriaPapp_A_ENG",
-                          no.progress=T)
+    matches <- getMatches(labbcat.url, pattern, participant.ids="UC427_ViktoriaPapp_A_ENG")
     
     ## check dataframe columns
     expect_true(length(matches$MatchId) >= 140)
@@ -206,12 +205,12 @@ test_that("getMatches includes segment info when segments layer searched", {
     expect_true(length(matches$Line)>= 140)
     expect_true(length(matches$LineEnd)>= 140)
     expect_true(length(matches$Text)>= 140)
-    expect_true(length(matches$Target.transcript)>= 140)
-    expect_true(length(matches$Target.transcript.start)>= 140)
-    expect_true(length(matches$Target.transcript.end)>= 140)
-    expect_true(length(matches$Target.segments)>= 140)
-    expect_true(length(matches$Target.segments.start)>= 140)
-    expect_true(length(matches$Target.segments.end)>= 140)
+    expect_true(length(matches$Target.word)>= 140)
+    expect_true(length(matches$Target.word.start)>= 140)
+    expect_true(length(matches$Target.word.end)>= 140)
+    expect_true(length(matches$Target.segment)>= 140)
+    expect_true(length(matches$Target.segment.start)>= 140)
+    expect_true(length(matches$Target.segment.end)>= 140)
 
     ## check dataframe column types
     expect_false(is.numeric(matches$MatchId))
@@ -219,9 +218,9 @@ test_that("getMatches includes segment info when segments layer searched", {
     expect_true(is.numeric(matches$Line))
     expect_true(is.numeric(matches$LineEnd))
     expect_false(is.numeric(matches$Text))
-    expect_false(is.numeric(matches$Target.transcript))
-    expect_true(is.numeric(matches$Target.transcript.start))
-    expect_true(is.numeric(matches$Target.transcript.end))
+    expect_false(is.numeric(matches$Target.word))
+    expect_true(is.numeric(matches$Target.word.start))
+    expect_true(is.numeric(matches$Target.word.end))
 })
 
 test_that("filter parameters of getMatches work", {
@@ -229,10 +228,10 @@ test_that("filter parameters of getMatches work", {
     if (!is.null(labbcatCredentials(labbcat.url, "demo", "demo"))) skip("Server not available")
 
     ## get matches
-    matchesAll <- getMatches(labbcat.url, list(orthography="the"), main.participant=F, no.progress=T)
-    matchesOnePerTranscript <- getMatches(labbcat.url, list(orthography="the"), matches.per.transcript=2, no.progress=T)
-    matchesMainParticipant <- getMatches(labbcat.url, list(orthography="the"), main.participant=T, no.progress=T)
-    matchesOneParticipant <- getMatches(labbcat.url, list(orthography="the"), participant.ids="UC427_ViktoriaPapp_A_ENG", no.progress=T)
+    matchesAll <- getMatches(labbcat.url, list(orthography="the"), main.participant=F)
+    matchesOnePerTranscript <- getMatches(labbcat.url, list(orthography="the"), matches.per.transcript=2)
+    matchesMainParticipant <- getMatches(labbcat.url, list(orthography="the"), main.participant=T)
+    matchesOneParticipant <- getMatches(labbcat.url, list(orthography="the"), participant.ids="UC427_ViktoriaPapp_A_ENG")
     
     ## check the number of results
     expect_true(length(matchesAll$MatchId) > length(matchesMainParticipant$MatchId))
@@ -251,10 +250,10 @@ test_that("words.context parameter of getMatches works", {
                      orthography = list(pattern = "knox")))))
 
     ## search for "knox", which has only one hit
-    noContext <- getMatches(labbcat.url, list(orthography="knox"), words.context=0, no.progress=T)
-    oneWord <- getMatches(labbcat.url, list(orthography="knox"), words.context=1, no.progress=T)
-    fiveWords <- getMatches(labbcat.url, list(orthography="knox"), words.context=5, no.progress=T)
-    wholeLine <- getMatches(labbcat.url, list(orthography="knox"), words.context=-1, no.progress=T)
+    noContext <- getMatches(labbcat.url, list(orthography="knox"), words.context=0)
+    oneWord <- getMatches(labbcat.url, list(orthography="knox"), words.context=1)
+    fiveWords <- getMatches(labbcat.url, list(orthography="knox"), words.context=5)
+    wholeLine <- getMatches(labbcat.url, list(orthography="knox"), words.context=-1)
 
     ## check the amount of context
     expect_true(is.na(noContext$Before.Match)[[1]])
@@ -282,10 +281,22 @@ test_that("getMatches pagination works", {
     pattern <- list(
         columns = list(
             list(layers = list(
-                     segments = list(pattern = "I")))))
+                     segment = list(pattern = "I")))))
     
     ## get matches - the total (as above) is >= 140, but we ask for the first 5
-    matches <- getMatches(labbcat.url, pattern, max.matches=5, no.progress=T)
+    matches <- getMatches(labbcat.url, pattern, max.matches=5)
     
     expect_equal(length(matches$MatchId), 5)
+})
+
+test_that("overlap.threshold parameter of getMatches works", {
+    skip_on_cran() # don't run tests that depend on external resource on CRAN
+    if (!is.null(labbcatCredentials(labbcat.url, "demo", "demo"))) skip("Server not available")
+
+    ## search for "mmm", which frequently appears in overlapping speech
+    allUtterances <- getMatches(labbcat.url, list(orthography="mmm"))
+    noOverlap <- getMatches(labbcat.url, list(orthography="mmm"), overlap.threshold=5)
+
+    ## allUtterances should be bigger than noOverlap
+    expect_true(nrow(allUtterances) > nrow(noOverlap))
 })